aro-partner-agent/pyproject.toml
[project]
name = "aro-partner-agent"
version = "0.1.0"
description = "Standalone ARO support agent — demonstrates MCP-powered remote partner agent"
authors = [
    {name = "Partner Agent Team", email = "team@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0,<0.117.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "openai>=1.30.0",
    "pyyaml>=6.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=5.13.0",
    "mypy>=1.17.1",
    "types-PyYAML>=6.0.12.20250915",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aro_agent"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=5.13.0",
    "mypy>=1.17.1",
    "types-PyYAML>=6.0.12.20250915",
    "aiohttp>=3.13.5",
    "pytest-aiohttp>=1.1.0",
]
